+++ src/math/ft_factorial.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_factorial.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lgoddijn <lgoddijn@student.codam.nl>       +#+  +:+       +#+        */
+/*   By: lgoddijn <lgoddijn@student.codam.nl >      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/28 17:22:10 by lgoddijn          #+#    #+#             */
-/*   Updated: 2024/01/01 15:01:38 by lgoddijn         ###   ########.fr       */
+/*   Updated: 2024/04/14 19:07:09 by lgoddijn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/ft_math.h"
 
-unsigned long long	ft_factorial(int n)
+uint64_t	ft_factorial(int32_t n)
 {
-	unsigned long long	result;
-	int					i;
+	uint64_t	result;
+	int32_t		i;
 
 	if (n < 0)
 		return (0);
